Apply refactoring
