+++ roguelike-game/src/main/java/config/configuration.yaml
@@ -1,5 +1,8 @@
 screen:
   width: 100
   height: 48
-map_generation:
+level_generation:
   predefined_to_generated_ratio: 0.5
+map_generation:
+  minumum_levels_count: 5
+  maximum_levels_count: 15
