+++ roguelike-game/src/test/java/models/MobTest.java
@@ -3,7 +3,6 @@ package test.models;
 import java.util.Random;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.BeforeEach;
 import static org.junit.jupiter.api.Assertions.*;
 
 import models.Mob;
